+++ App.tsx
@@ -2,23 +2,23 @@
 import { StyleSheet, Text, View, Image, ScrollView } from 'react-native';
 import Splash from './screens/splashScreen';
 import OnboardingScreen from './screens/OnboardingScreen';
-import Home from './screens/homeScreen';
-import Explore from './screens/Explorescreen';
-import History from './screens/Historyscreen';
-import Profile from './screens/profilescreen';
-import DetailCategory from './screens/DetailCategoryScreen';
-import DetailParking from './screens/DetailParkingScreen';
-import Notification from './screens/NotificationScreen';
-import DetailHistory from './screens/DetailHistory';
-import EditProfile from './screens/EditProfileScreen';
-import ParkingRoute from './screens/parkingRouteScreen';
+import Home from './screens/home/homeScreen';
+import Explore from './screens/parking/Explorescreen';
+import History from './screens/parking/Historyscreen';
+import Profile from './screens/profile/profilescreen';
+import DetailCategory from './screens/parking/DetailCategoryScreen';
+import DetailParking from './screens/parking/DetailParkingScreen';
+import Notification from './screens/notification/NotificationScreen';
+import DetailHistory from './screens/parking/DetailHistory';
+import EditProfile from './screens/profile/EditProfileScreen';
+import ParkingRoute from './screens/parking/parkingRouteScreen';
 import Onboarding from './screens/OnboardingScreen';
-import TrackingPark from './screens/TrackingPark';
-import UpgradePro from './screens/UpgradePro';
-import ChoosePlan from './screens/Chooseplan';
-import Payment from './screens/Payment';
-import Details from './screens/PaymentDetails';
-import Success from './screens/PaymentSuccess';
+import TrackingPark from './screens/parking/TrackingPark';
+import UpgradePro from './screens/plan/UpgradePro';
+import ChoosePlan from './screens/plan/Chooseplan';
+import Payment from './screens/payment/Payment';
+import Details from './screens/payment/PaymentDetails';
+import Success from './screens/payment/PaymentSuccess';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import {BookDetailScreen, BookScreen, ChoosePlanScreen,
@@ -26,19 +26,19 @@ ChooseSpaceScreen, DetailCategoryScreen, DetailHistoryScreen,
 EditProfileScreen, ExploreScreen, HistoryScreen, HomeScreen,
 NotificationScreen, ParkingRouteScreen,  PaymentDetailsScreen, 
 PaymentScreen, PaymentSuccessScreen, ProfileScreen, ShowMapsScreen,
-SignUpScreen, TrackingParkScreen, UpgradeProScreen} from './screens/Navigator';
-import { LoginScreen } from './screens/Navigator';
-import { RequestCodeScreen } from './screens/Navigator';
-import { RequestOtpScreen } from './screens/Navigator';
-import { ForgotPasswordScreen } from './screens/Navigator';
-import { ConfirmPasswordScreen } from './screens/Navigator';
-import { PhoneLoginScreen } from './screens/Navigator';
-import DetailParkingScreen from './screens/DetailParkingScreen';
-import ShowMaps from './screens/ShowMaps';
+SignUpScreen, TrackingParkScreen, UpgradeProScreen} from './screens/navigator/Navigator';
+import { LoginScreen } from './screens/navigator/Navigator';
+import { RequestCodeScreen } from './screens/navigator/Navigator';
+import { RequestOtpScreen } from './screens/navigator/Navigator';
+import { ForgotPasswordScreen } from './screens/navigator/Navigator';
+import { ConfirmPasswordScreen } from './screens/navigator/Navigator';
+import { PhoneLoginScreen } from './screens/navigator/Navigator';
+import DetailParkingScreen from './screens/parking/DetailParkingScreen';
+import ShowMaps from './screens/parking/ShowMaps';
 import 'react-native-gesture-handler';
-import Book from './screens/Book';
-import BookDetail from './screens/BookDetail';
-import ChooseSpace from './screens/ChooseSpace';
+import Book from './screens/book/Book';
+import BookDetail from './screens/book/BookDetail';
+import ChooseSpace from './screens/parking/ChooseSpace';
 import { UserProvider } from './contexts/userContext';
