+++ screens/navigator/Navigator.tsx
@@ -1,34 +1,34 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import Onboarding from './OnboardingScreen';
-import Login from './logins/login';
-import Register from './logins/signup';
-import PhoneLogin from './logins/LoginWithPhone';
-import RequestOTP from './logins/forgotpassword/Request-otp'; 
-import ConfirmForgotPassword from './logins/forgotpassword/confirmation'; 
-import RequestCode from './logins/forgotpassword/Request-code';
-import History from './Historyscreen';
-import EditProfile from './EditProfileScreen';
-import Explore from './Explorescreen';
-import UpgradePro from './UpgradePro';
-import ChoosePlan from './Chooseplan';
-import ParkingRoute from './parkingRouteScreen';
-import TrackingPark from './TrackingPark';
-import Notification from './NotificationScreen';
-import Profile from './profilescreen';
-import DetailCategory from './DetailCategoryScreen';
-import DetailHistory from './DetailHistory';
-import DetailParking from './DetailParkingScreen';
+import Onboarding from '../OnboardingScreen';
+import Login from '../logins/login';
+import Register from '../logins/signup';
+import PhoneLogin from '../logins/LoginWithPhone';
+import RequestOTP from '../logins/forgotpassword/Request-otp'; 
+import ConfirmForgotPassword from '../logins/forgotpassword/confirmation'; 
+import RequestCode from '../logins/forgotpassword/Request-code';
+import History from '../parking/Historyscreen';
+import EditProfile from '../profile/EditProfileScreen';
+import Explore from '../parking/Explorescreen';
+import UpgradePro from '../plan/UpgradePro';
+import ChoosePlan from '../plan/Chooseplan';
+import ParkingRoute from '../parking/parkingRouteScreen';
+import TrackingPark from '../parking/TrackingPark';
+import Notification from '../notification/NotificationScreen';
+import Profile from '../profile/profilescreen';
+import DetailCategory from '../parking/DetailCategoryScreen';
+import DetailHistory from '../parking/DetailHistory';
+import DetailParking from '../parking/DetailParkingScreen';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
-import ForgotPassword from './logins/forgotpassword/forgot';
-import Home from './homeScreen';
-import Payment from './Payment';
-import Details from './PaymentDetails';
-import PaymentSuccess from './PaymentSuccess';
-import ShowMaps from './ShowMaps';
-import Book from './Book';
-import BookDetail from './BookDetail';
-import ChooseSpace from './ChooseSpace';
+import ForgotPassword from '../logins/forgotpassword/forgot';
+import Home from '../home/homeScreen';
+import Payment from '../payment/Payment';
+import Details from '../payment/PaymentDetails';
+import PaymentSuccess from '../payment/PaymentSuccess';
+import ShowMaps from '../parking/ShowMaps';
+import Book from '../book/Book';
+import BookDetail from '../book/BookDetail';
+import ChooseSpace from '../parking/ChooseSpace';
 
 
 const Stack = createStackNavigator();
